+++ sccp/figures/figure5.py
@@ -2,11 +2,12 @@
 Parafac2 implementation on PBMCs treated wtih PopAlign/Thompson drugs: investigating UMAP
 """
 import numpy as np
-import seaborn as sns
-import pandas as pd
 from .common import (
     subplotLabel,
     getSetup,
+    flattenData,
+    plotDrugUMAP,
+    plotGeneUMAP
 )
 from ..imports.scRNA import ThompsonXA_SCGenes
 from ..parafac2 import parafac2_nd
@@ -27,7 +28,7 @@ def makeFigure():
 
     _, factors, projs, _ = parafac2_nd(
         data,
-        rank=13,
+        rank=30,
     )
 
     dataDF, projDF = flattenData(data, factors, projs)
@@ -47,50 +48,4 @@ def makeFigure():
 
     return f
 
-def flattenData(data, factors, projs):
-    cellCount = []
-    for i in range(factors[0].shape[0]):
-        cellCount = np.append(cellCount, projs[i].shape[0])
-    
-    flatProjs = np.empty([int(np.sum(cellCount)), projs[0].shape[1]])
-    flatData = np.empty([int(np.sum(cellCount)), len(data.variable_labels)])
-    cellStart = [0]; drugNames = []
-    
-    for i in range(factors[0].shape[0]):
-        cellStart = np.append(cellStart, cellStart[i] + cellCount[i])
-        flatProjs[int(cellStart[i]): int(cellStart[i+1])] = projs[i]
-        flatData[int(cellStart[i]): int(cellStart[i+1])] = data.X_list[i]
-        drugNames = np.append(drugNames, np.repeat(data.condition_labels[i], cellCount[i]))
-    
-    cmpNames = [f"Cmp. {i}" for i in np.arange(1, factors[0].shape[1] + 1)]
-    projDF = pd.DataFrame(data=flatProjs, columns = cmpNames)
-    dataDF = pd.DataFrame(data=flatData, columns = data.variable_labels)
-    projDF["Drug"] = drugNames
-    dataDF["Drug"] = drugNames
-    
-    return dataDF, projDF
-
-def plotGeneUMAP(genes, umapPoints, dataDF, axs):
-    for i, genez in enumerate(genes):
-        geneList = dataDF[genez].to_numpy()
-        umapDF = pd.DataFrame({"UMAP1": umapPoints[::20, 0],
-                "UMAP2": umapPoints[::20, 1],
-                genez: geneList[::20],
-            })
-        sns.scatterplot(data=umapDF, x="UMAP1", y="UMAP2", hue=genez, s=5, ax=axs[i])
-        axs[i].set(xlim=(-10, 20), ylim=(-15, 20))
-        
-    return 
-
-def plotDrugUMAP(drugs, totaldrugs, umapPoints, axs):
-    for i, drugz in enumerate(drugs):
-        drugList = np.asarray(totaldrugs == drugz).astype(int)
-        umapDF = pd.DataFrame({"UMAP1": umapPoints[::20, 0],
-                "UMAP2": umapPoints[::20, 1],
-                drugz: drugList[::20],
-            })
-        sns.scatterplot(data=umapDF, x="UMAP1", y="UMAP2", hue=drugz, s=5,  palette="muted", ax=axs[i])
-        
-    return
-
